refactor(oid4vc): split public routes into modules and update tests
